+++ base/src/main/java/org/aya/core/meta/Meta.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.core;
+package org.aya.core.meta;
 
 import kala.collection.SeqView;
 import kala.collection.immutable.ImmutableSeq;
@@ -15,7 +15,6 @@ import org.aya.ref.AnyVar;
 import org.aya.util.Arg;
 import org.aya.util.error.SourcePos;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 /**
  * @author ice1000, re-xyr
@@ -26,13 +25,8 @@ public final class Meta implements AnyVar {
   public final @NotNull ImmutableSeq<Term.Param> contextTele;
   public final @NotNull ImmutableSeq<Term.Param> telescope;
   public final @NotNull String name;
-  public final @Nullable Term result;
+  public final @NotNull MetaInfo info;
   public final @NotNull SourcePos sourcePos;
-  /**
-   * Usually implies {@link Meta#result} == null,
-   * asserts that the current meta is a type meta.
-   */
-  public final boolean isType;
   public final @NotNull MutableList<Tuple2<Subst, Term>> conditions = MutableList.create();
 
   public SeqView<Term.Param> fullTelescope() {
@@ -42,14 +36,13 @@ public final class Meta implements AnyVar {
   private Meta(
     @NotNull ImmutableSeq<Term.Param> contextTele,
     @NotNull ImmutableSeq<Term.Param> telescope,
-    @NotNull String name, @Nullable Term result,
-    boolean isType, @NotNull SourcePos sourcePos
+    @NotNull String name, @NotNull MetaInfo info,
+    @NotNull SourcePos sourcePos
   ) {
     this.contextTele = contextTele;
     this.telescope = telescope;
     this.name = name;
-    this.result = result;
-    this.isType = isType;
+    this.info = info;
     this.sourcePos = sourcePos;
   }
 
@@ -57,7 +50,7 @@ public final class Meta implements AnyVar {
     @NotNull ImmutableSeq<Term.Param> contextTele, @NotNull String name,
     @NotNull SourcePos sourcePos
   ) {
-    return new Meta(contextTele, ImmutableSeq.empty(), name, null, true, sourcePos);
+    return new Meta(contextTele, ImmutableSeq.empty(), name, new MetaInfo.AnyType(), sourcePos);
   }
 
   public static @NotNull Meta from(
@@ -66,9 +59,9 @@ public final class Meta implements AnyVar {
   ) {
     if (result instanceof PiTerm pi) {
       var buf = MutableList.<Term.Param>create();
-      var r = pi.parameters(buf);
-      return new Meta(contextTele, buf.toImmutableSeq(), name, r, false, sourcePos);
-    } else return new Meta(contextTele, ImmutableSeq.empty(), name, result, false, sourcePos);
+      var r = new MetaInfo.Result(pi.parameters(buf));
+      return new Meta(contextTele, buf.toImmutableSeq(), name, r, sourcePos);
+    } else return new Meta(contextTele, ImmutableSeq.empty(), name, new MetaInfo.Result(result), sourcePos);
   }
 
   public @NotNull PiTerm asPi(
@@ -76,12 +69,24 @@ public final class Meta implements AnyVar {
     @NotNull ImmutableSeq<Arg<Term>> contextArgs
   ) {
     assert telescope.isEmpty();
-    var domVar = new Meta(contextTele, ImmutableSeq.empty(), domName, result, isType, sourcePos);
-    var codVar = new Meta(contextTele, ImmutableSeq.empty(), codName, result, isType, sourcePos);
+    // TODO[isType]: this one should be piDom
+    var domVar = new Meta(contextTele, ImmutableSeq.empty(), domName, info, sourcePos);
+    // TODO[isType]: this one should be piCod
+    var codVar = new Meta(contextTele, ImmutableSeq.empty(), codName, info, sourcePos);
     var dom = new MetaTerm(domVar, contextArgs, ImmutableSeq.empty());
     var cod = new MetaTerm(codVar, contextArgs, ImmutableSeq.empty());
     var domParam = new Term.Param(Constants.randomlyNamed(sourcePos), dom, explicit);
     return new PiTerm(domParam, cod);
+  }
+
+  public @NotNull MetaTerm asPiDom(
+    @NotNull Term resultNew, @NotNull ImmutableSeq<Arg<Term>> contextArgs
+  ) {
+    assert telescope.isEmpty();
+    assert info instanceof MetaInfo.AnyType;
+    // TODO[isType]: this one should be piDom
+    var typed = new Meta(contextTele, ImmutableSeq.empty(), name, info, sourcePos);
+    return new MetaTerm(typed, contextArgs, ImmutableSeq.empty());
   }
 
   @Override public @NotNull String name() {
